move setContextCurrentValue into reconcilation module
